Week-5/requirements.txt
# LangChain & LangGraph
langchain>=0.1.0
langchain-aws>=0.2.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.2.0

# AWS Bedrock
boto3>=1.34.0

# Web Search
ddgs>=7.0.0

# Utilities
python-dotenv>=1.0.0
